Create method to get player's next move
Create a method named nextMove() which asks the player to make their move.  They will be required to enter an integer number between 1 and 9, corresponding to the square number on the Tic Tac Toe grid.

This method will only accept a single integer input from the user.  Any other input will be rejected until the correct input is provided.